requirements.txt
alembic==1.1.0
appnope==0.1.0
backtrader
Babel==2.7.0 
backports.shutil-get-terminal-size==1.0.0
cement==3.0.4
certifi
chardet==3.0.4
colorama==0.4.1
contextlib2==0.5.5
configparser==3.8.1
Cython==0.29.13
decorator==4.4.0
Delorean==1.0.0
enum34==1.1.6
gryphon-cdecimal==2.3
gryphon-money==1.2.1
humanize==0.5.1
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
line-profiler==2.1.2
more-itertools==7.2.0
mysqlclient
nose==1.3.7
pandas
pathlib2==2.3.4
pexpect==4.7.0
phonenumbers
pickleshare==0.7.5
prompt-toolkit==2.0.9
prompter
ptyprocess==0.6.0
pycryptodome
Pygments==2.4.2
pylibmc==1.6.1
python-dateutil==2.8.0
python-dotenv
pytz==2019.2
raven==6.10.0
redis==3.3.8
rednose==1.3.0
requests==2.22.0
requests-futures==0.9.7
requests-toolbelt==0.8.0
retrying==1.3.3
rq
scandir==1.10
simplegeneric==0.8.1
six==1.12.0
SQLAlchemy==1.3.8
sure==1.4.11
TA-Lib
termcolor==1.1.0
termstyle==0.1.11
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
